+++ version_history.md
@@ -1,5 +1,14 @@
 #Version history#
 
+##1.2 (08/19/12)
+* Added <tt>_.selectWhile</tt>
+* Added <tt>_.mapReduce</tt> and <tt>_.mapr</tt> (as alias)
+* Added <tt>_.mapReduceRight</tt> and <tt>_.maprr</tt> (as alias)
+* Added <tt>_.bindn</tt>
+* Added <tt>_.appendLists</tt>
+* Updated Moses_Lib_Test.lua
+* Updated documentation
+
 ##1.1 (08/04/12)
 * Removed <tt>_.contains</tt> as alias to <tt>_.include</tt>
 * Added <tt>_.removeRange</tt> (as Array function)